Added back in the missing front case steel
As mkovari pointed out, the previous commits did not consider all of the front case steel that exists. This commit corrects for this extra steel, but in the front case vertical modulus, not the thickness. In this manner the extra steel can contribute to the axial dynamics without messing up the correct stress-concentrated and less-stiff toroidal behavior (coming from having points with very thin radial extent)