Changed properties to conductor and fixed a bug
Before, eyoung_winding and poisson_winding were called the properties of the winding pack, which was wrong. We want the properties of the conductor to be the input, hence eyoung_cond and poisson_cond.

Also fixed a bug in which the lateral case contribution to the vertical properties was incorrect.